requirements.txt
Django==5.0.1
mysqlclient==2.2.1
requests==2.31.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
sendgrid==6.11.0
